[DX] Extract BatchStatus and ExitStatus
