+++ spec/Pim/Bundle/ImportExportBundle/Normalizer/JobExecutionNormalizerSpec.php
@@ -4,7 +4,7 @@ namespace spec\Pim\Bundle\ImportExportBundle\Normalizer;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
-use Akeneo\Bundle\BatchBundle\Job\BatchStatus;
+use Akeneo\Component\Batch\Job\BatchStatus;
 use PhpSpec\ObjectBehavior;
 use Symfony\Component\Serializer\SerializerInterface;
 use Symfony\Component\Translation\TranslatorInterface;
