+++ src/Akeneo/Bundle/BatchBundle/Entity/JobExecution.php
@@ -3,8 +3,8 @@
 namespace Akeneo\Bundle\BatchBundle\Entity;
 
 use Akeneo\Bundle\BatchBundle\Item\ExecutionContext;
-use Akeneo\Bundle\BatchBundle\Job\BatchStatus;
-use Akeneo\Bundle\BatchBundle\Job\ExitStatus;
+use Akeneo\Component\Batch\Job\BatchStatus;
+use Akeneo\Component\Batch\Job\ExitStatus;
 use Akeneo\Bundle\BatchBundle\Job\RuntimeErrorException;
 use Doctrine\Common\Collections\ArrayCollection;
 // TODO extract mapping in yml file
