+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/Entity/StepExecutionTest.php
@@ -5,8 +5,8 @@ namespace Akeneo\Bundle\BatchBundle\Tests\Unit\Entity;
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
 use Akeneo\Bundle\BatchBundle\Item\ExecutionContext;
-use Akeneo\Bundle\BatchBundle\Job\BatchStatus;
-use Akeneo\Bundle\BatchBundle\Job\ExitStatus;
+use Akeneo\Component\Batch\Job\BatchStatus;
+use Akeneo\Component\Batch\Job\ExitStatus;
 
 /**
  * Test related class
