+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/Job/BatchStatusTest.php
@@ -2,7 +2,7 @@
 
 namespace Akeneo\Bundle\BatchBundle\Tests\Unit\Job;
 
-use Akeneo\Bundle\BatchBundle\Job\BatchStatus;
+use Akeneo\Component\Batch\Job\BatchStatus;
 
 /**
  * Tests related to the BatchStatus class
