+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/Job/JobTest.php
@@ -4,8 +4,8 @@ namespace Akeneo\Bundle\BatchBundle\Tests\Unit\Job;
 
 use Akeneo\Bundle\BatchBundle\Entity\JobExecution;
 use Akeneo\Bundle\BatchBundle\Entity\JobInstance;
-use Akeneo\Bundle\BatchBundle\Job\BatchStatus;
-use Akeneo\Bundle\BatchBundle\Job\ExitStatus;
+use Akeneo\Component\Batch\Job\BatchStatus;
+use Akeneo\Component\Batch\Job\ExitStatus;
 use Akeneo\Bundle\BatchBundle\Job\Job;
 use Akeneo\Component\Batch\Step\ItemStep;
 use Akeneo\Bundle\BatchBundle\Tests\Unit\Step\IncompleteStep;
