+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/Manager/JobExecutionManagerTest.php
@@ -2,8 +2,8 @@
 
 namespace Akeneo\Bundle\BatchBundle\Tests\Unit\Manager;
 
-use Akeneo\Bundle\BatchBundle\Job\BatchStatus;
-use Akeneo\Bundle\BatchBundle\Job\ExitStatus;
+use Akeneo\Component\Batch\Job\BatchStatus;
+use Akeneo\Component\Batch\Job\ExitStatus;
 use Akeneo\Bundle\BatchBundle\Manager\JobExecutionManager;
 
 class JobExecutionManagerTest extends \PHPUnit_Framework_TestCase
