+++ src/Akeneo/Bundle/BatchBundle/Tests/Unit/Step/IncompleteStep.php
@@ -3,14 +3,14 @@
 namespace Akeneo\Bundle\BatchBundle\Tests\Unit\Step;
 
 use Akeneo\Bundle\BatchBundle\Entity\StepExecution;
-use Akeneo\Bundle\BatchBundle\Job\BatchStatus;
+use Akeneo\Component\Batch\Job\BatchStatus;
 use Akeneo\Component\Batch\Step\AbstractStep;
 
 /**
  * Step used for test and always declared a incomplete execution
  *
  */
-class IncompleteStep extends \Akeneo\Component\Batch\Step\AbstractStep
+class IncompleteStep extends AbstractStep
 {
     /**
      * {@inheritdoc}
