+++ src/Pim/Component/Connector/Step/TaskletInterface.php
@@ -11,7 +11,7 @@ use Akeneo\Component\Batch\Step\StepExecutionAwareInterface;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-interface TaskletInterface extends \Akeneo\Component\Batch\Step\StepExecutionAwareInterface
+interface TaskletInterface extends StepExecutionAwareInterface
 {
     /**
      * Execute the tasklet
