+++ src/Pim/Component/Connector/Step/TaskletStep.php
@@ -11,7 +11,7 @@ use Pim\Component\Connector\Step\TaskletInterface;
  * @copyright 2015 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class TaskletStep extends \Akeneo\Component\Batch\Step\AbstractStep
+class TaskletStep extends AbstractStep
 {
     /** @var array */
     protected $configuration;
